Divide larger modules into sensible new modules. Start adding auto parsing of ReproIn-style series description.
