+++ setup.py
@@ -45,7 +45,7 @@ setup(
     # For a discussion on single-sourcing the version across setup.py and the
     # project code, see
     # https://packaging.python.org/en/latest/single_source_version.html
-    version='2022.2.28',  # Required
+    version='2022.5.9',  # Required
 
     # This is a one-line description or tagline of what your project does. This
     # corresponds to the "Summary" metadata field:
@@ -145,7 +145,7 @@ setup(
     # For an analysis of "install_requires" vs pip's requirements files see:
     # https://packaging.python.org/en/latest/requirements.html
     install_requires=['pydicom>=2.2',
-                      'pybids==0.12.4',
+                      'pybids>=0.15',
                       'numpy>=1.21'],
 
     # List additional groups of dependencies here (e.g. development
